app/requirements.txt
fastapi
uvicorn
python-multipart
requests
py-zerox
greenlet
pydantic
sqlalchemy
asyncpg
pydantic-settings
pydantic-core
pydantic-settings
google-cloud-storage